Create stuff2.c
wrote matrixMultSingle() which, with some modification will run a multithreaded matrix multiplication on a single matrix. Also I moved all code to print a matrix to be self reliant. We can now call printMatrixMult() after a call to multAll() or matrixMultSingle() to check the results. This should now compile with the Cuda compiler.